requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
openai>=1.0.0
pyyaml>=6.0
python-dotenv>=0.19.0

# Data analysis and visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
scipy>=1.9.0

# Jupyter and interactive analysis
jupyter>=1.0.0
jupyterlab>=3.0.0
ipywidgets>=7.6.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Progress bars and utilities
tqdm>=4.64.0
click>=8.0.0

# Optional: For advanced analysis
scikit-learn>=1.1.0
statsmodels>=0.13.0
